requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
black==23.7.0
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.6 ; python_version >= '3.7'
django==4.2.3
django-cors-headers==4.2.0
git+https://github.com/sageteam-org/django-iranian-cities.git@b4d5f63a19acb8555e54b9116c54cb3e1723c7e9#egg=django-iranian-cities
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.5.0
filetype==1.2.0
idna==3.4 ; python_version >= '3.5'
mypy-extensions==1.0.0 ; python_version >= '3.5'
packaging==23.1 ; python_version >= '3.7'
pathspec==0.11.1 ; python_version >= '3.7'
pillow==10.0.0
platformdirs==3.9.1 ; python_version >= '3.7'
pyjwt==2.8.0 ; python_version >= '3.7'
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4 ; python_version >= '3.5'
urllib3==2.0.4 ; python_version >= '3.7'
